Fix usage of JRuby detection methods
The `Appsignal::System.jruby?` method is used by the library itself, and
to test values from the library.

The `DependencyHelper.running_jruby?` method is used to determine if
certain blocks of specs need to be run or not.

These two methods have different implementations. I'm not sure why. I
think the DependencyHelper's implementation is better.

[skip changeset] because it's a small internal test suite refactor.