Add sequel-rails event formatter for SQL queries
The sequel-rails gem has its own ActiveSupport::Notifications
instrumentor, tracking events about Sequel queries that are being made.
https://github.com/TalentBox/sequel-rails/blob/db6778f1fbdc15a4a53f9f406dc6354debdd17fb/lib/sequel_rails/sequel/database/active_support_notification.rb#L11-L16

This conflicts with our own sequel instrumentation, recording a
duplicate event. This duplicate event can be worked around by setting
`instrument_sequel` to `false`, disabling the AppSignal integration.
This issue is not addressed in this commit. It requires the manual
config change.

The problem then is that the sequel-rails ActiveSupport::Notifications
event was not handled in any special way, so the event was stored
without an event body, not showing the query that was executed.

To track the SQL query that was made, add an AppSignal event formatter
(used by the ActiveSupport::Notifications integration) to extract the
SQL query from the event and store it on the AppSignal event.

Fixes #799