Add send_session_data option
This replaces the `skip_session_data` option. This option is now
deprecated.

As discussed in https://github.com/appsignal/integration-guide/pull/114,
the `skip_session_data` is inconsistently named compared to our other
options that do the same thing, but start with `send_`.

The behavior of `send_session_data` is also the exact opposite of
`skip_session_data`. It's on by default and if turned off it won't send
the session data.

Only when `send_session_data` is not set by the user, will
`skip_session_data` be listened to. If `send_session_data`'s value is
determined by the `maintain_backwards_compatibility` method, its value
is set in the `system` source, because the system figured out what value
it needs to be set to.  If the user has manually configured it, it will
not set it in the system source.

In the future, when the `skip_session_data` option is removed, the
`send_session_data` option should be added to the `DEFAULT_CONFIG`
constant in the `Config` class and this behavior should all be removed.