Fix deprecation warning for working_dir_path
It would only print the warning if the config option was used in the
file source, but if the `APPSIGNAL_WORKING_DIR_PATH` environment
variable was set, it would not print the same deprecation warning. This
only makes the deprecation more visible. This is necessary before we
fully remove it in the next major version.

I've added a changeset, because it prints the warning in more scenarios
and those are scenarios we want to inform our users about before
removing the config option entirely.

Fixes https://github.com/appsignal/appsignal-ruby/issues/805