Add config override source (#821)
Add the `override` config source to track the final decisions made by
the Ruby gem concerning some config options.

Currently only the `send_session_data` option is tracked this way. When
the deprecated `skip_session_data` option is set, but
`send_session_data` is not, it defaults to the `skip_session_data`
value. This decision was previously stored on the `system` source.

Now it's more clear that the Ruby gem made this particular decision and
we don't have to figure it out in the report viewer.

Fixes #811