Clean ActiveRecord::RecordNotUnique error message
Like we've done before for `PG::UniqueViolation`, sanitize the
`ActiveRecord::RecordNotUnique` message.

The `RecordNotUnique` error class is a type of error that wraps around
errors from database adapter gems like the "pg" gem's
`PG::UniqueViolation`. The message for this `RecordNotUnique` error is
the same for `PG::UniqueViolation` in this case, prefixed with error
class name.

Fixes  #833