Fix Rails 6.1.5 ActiveJob build errors
For whatever reason, Rails 6.1.5 reverts to the behaviour from
versions previous to Rails 6.1, when sending Ruby hashes through
ActiveJob in Ruby 3.1.

This means that, in the JSON sent internally by ActiveJob, the hash
is no longer wrapped in an additional JSON object to indicate that
the hash has the `ruby2_keywords_hash?` flag activated.