Change build matrix from denylist to allowlist
The "only" config will allow you to set which versions it should run for
rather than what versions it shouldn't run for. Over time, as packages
no longer get updated, or we tests a specific older versions, we need to
keep adding Ruby versions to the exclude config, rather than just
configuring which Ruby versions it should run for.

This should be slightly easier to maintain and easier to read in the
`build_matrix.yml` file.

The "exclude" config still exists, but cannot be used in combination
with the "only" config.