Use Dir.pwd instead of ENV['PWD'] in Capistrano integration (#848)
* Use Dir.pwd instead of ENV['PWD']

I'm using a pretty different setup for my Deploy with Capistrano, which
includes deploying from a dev container in rootless podman, using the
Ruby on Whales guide, with some changes.

I don't know exactly why it happens, but when deploying ENV["PWD"]
returns nil. I tinkered a little and it seems that Dir.pwd is more
reliable.

* Use Dir.chdir to mock directory in Capistrano spec

Instead of mocking the `Dir.pwd` method, switch the current directory to
what the test expects it to be. This doesn't interfere with our internal
test system.

Co-authored-by: Tomás Coêlho <tomascoelho6@gmail.com>
Co-authored-by: Tom de Bruijn <tom@tomdebruijn.com>