MRI VM Stats (#852)
* Add Ruby 3.2 preview to build matrix

Add the preview version of Ruby 3.2 to the build matrix. Exclude
testing against gemsets that were already excluded from Ruby 3.1.

If the requested Ruby version cannot be found, fetch the latest
build definitions for `ruby-build` from its Git repository and
attempt to build from source. Source builds are slow, so they're
cached in Semaphore.

* Test against the latest grape gem versions

The version was restricted to 0.14.0, which was released more than
five years ago. Test against the latest versions of the library.

* MriProbe tracks thread counts

Take Thread.list.size and set it as a gauge called thread_count.

* Switch MriSpec to use a mock to test sent values

Instead of using the receive function that mocks (noun) the Appsignal
class, use a mock (verb) to test the sent values.

This switch allows for checking multiple subsequent calls to
add_distribution_value and set_gauge.

* MriProbe tracks GC runs

Use GC.count, and reports that as "gc_runs"

* Check keys and values for MRI data

Instead of just making sure the correct metrics are reported, add extra
checks to make sure the keys are present and the value is non-nil.

* Add before block to MRI stats tests

Instead of calling probe.call in every test run, move it to a before
block, which is equivalent.

* Add GC stats

Call GC.stat, take :total_allocated_objects, :major_gc_count,
:minor_gc_count, :heap_live and :heap_free

* Split gc_stats up into different metrics

Add a separate gauge for total_allocated_objects, and split
major/minor_gc_count and heap_live/heap_free into their own
distributions.

* Add gc_total_time to MRI stats

Use Appsignal::GarbageCollectionProfiler to fetch the number of garbage
collection runs, and report it as gc_total_time.

* Fix Rubocop warnings

Disable Naming/AccessorMethodName for AppsignalMock.set_gauge, as
set_gauge is the name of the faked method.

* Fall back to total_allocated_object in MRI stats

In Ruby 2.1.x, the :total_allocated_objects key in GC.stats was called
:total_allocated_object (note the missing "s"). Fall down to
:total_allocated_object if :total_allocated_objects is nil.

* Fall back to heap_(live|free)_slot in MRI stats

In Ruby 2.1.x, the :heap_(live|free)_slots key in GC.stats was called
:heap_(live|free)_slot (note the missing "s"). Fall down to
:heap_(live|free)_slot if :heap_(live|free)_slots is nil.

* Use RubyVM.stat[:constant_cache] on Ruby 3.2

Ruby 3.2 adds :constant_cache and drops :global_constant_state. Check
:constant_cache first, and sum its values if it exists, otherwise fall
back to :global_constant_state.

* Add changeset for VM Dashboard addition

Marked this change as a minor release, as this is an added feature that
warrants a blogpost.

* Clarify MRI probe changeset

Update .changesets/add-a-mri-vm-probe-to-add-a-magic-dashboard.md

Co-authored-by: Noemi <45180344+unflxw@users.noreply.github.com>

Co-authored-by: Noemi <noemi@appsignal.com>
Co-authored-by: Noemi <45180344+unflxw@users.noreply.github.com>