Store delta gauge of allocated objects in Ruby
The total_allocated_objects metric will continuously report an growing
value, which resets on every restart or new deploy. Creating a lot of
highs and lows.

This makes it difficult to track if the app is allocating more or less
objects between deploys, other than tracking how steep the increase is.

The delta of this metric reports the difference between probe runs,
which should result in a more stable graph in apps with stable memory
allocation.