Store delta gauge of Garbage Collection counts
The `gc_count` metric will continuously report an growing
value, which resets on every restart or new deploy. Creating a lot of
highs and lows.

This makes it difficult to track if the app is doing more or less
garbage collection between deploys, other than tracking how steep the
increase is visually.

The delta of this metric reports the difference between probe runs,
which should result in a more stable graph that shows more heavy garbage
collection as a spike.