Store delta gauge of Garbage Collection total time
The `gc_total_time` metric will continuously report an growing value,
which resets on every restart, new deploy, or every 23 days (see the
`Appsignal::GarbageCollectionProfiler#total_time` method). Creating a
lot of highs and lows in a graph.

This makes it difficult to track if the app is taking longer for garbage
collection between deploys, other than tracking how steep the increase
is visually.

The delta of this metric reports the difference between probe runs,
which should result in a more stable graph that shows longer garbage
collection as a spike.