Change `gauge_delta` so it expects a block
The `gauge_delta` helper would return the delta between the value
passed in the current call and the value stored in its cache from
a previous call, or `nil` if no previous call had taken place yet.

Because of this, at every point where `gauge_delta` is used, there
is an `if` that follows, that checks for the value not being `nil`.

This commit changes the `gauge_delta` method so that it expects a
block that will take the delta value as argument. This block will
not be called on the first invocation of `gauge_delta` for a given
cache key, avoiding the need to manually handle the case where the
delta is `nil`.