Add queue_start test helper to fetch queue start
Previously we asserted a method call and didn't check the value it was
called with. I've adapted some implementation of the queue start
assertion I found in the Transaction specs to create a test for the
Rails instrumentation.

This should really be fixed in
https://github.com/appsignal/appsignal-agent/issues/293 (no hurry), but
at least we now has a somewhat smaller helper we can reuse in the test
suite.

We can also start using this new method of asserting queue time in other
tests, but I wanted to finish this quickly and not get hung up on
updating everything. We can do this at a later time.