Fix Rails error on unknown request method
If `ActionDispatch::Request#request_method` is called when the request
method is not an known request method by Rails it will raise an
`ActionController::UnknownHttpMethod`.

We do not expect this method to raise an error, so it broke our Rails
instrumentation, didn't close the transaction which would cause it be
reused by the next request handled by Rails.

Add exception handling to the method calls so it doesn't break our
instrumentation anymore.

Fixes #851
Fixes https://github.com/appsignal/support/issues/201