Remove return false from void methods
The `Appsignal.add_breadcrumb` and `Appsignal.tag_request` methods are
defined to return a void, nothing. No one should think this returns any
value or base their app behavior on it. If something is returned by
being the last value of the method that should also not be relied upon
as it may change without notice.

Remove the return value of `false` for these two methods.