Add Transaction.current? helper
This helper returns true if a transaction is active and false if there
is none. This makes the following cumbersome code easier:

```ruby
# Before
current = Appsignal::Transaction.current
if !current.nil? && !current.nil_transaction?
  current.set_error(error)
end

# After
if Appsignal::Transaction.current?
  Appsignal::Transaction.current.set_error(error)
end
```

This is usually not necessary when no transaction is active, because the
`NilTransaction` object returned by `Transaction.current` will trigger
any errors regardless of what you call on it, but sometimes you need to
know if there really is a transaction active or not.

We do this internally in some helpers to avoid overhead fetching data
when no transaction is active. This can also be useful for users
implementing their own instrumentation, so I've documented the
`Transaction.current` and `Transaction.current?` methods.

I've also updated calls to `Transaction.current.nil?` because they do
not check if the returned transaction is a `NilTransaction`. You need to
use `NilTransaction.nil_transaction?` for that.