Remove GC::Profiler.enable
I don't think we should provide this in our gem, but have users add this
to their own app manually in a Rails's `application.rb` file,
`development.rb` file or deeper in their app code.

This way they also have more control over when to disable it again, by
using `GC::Profiler.disable` if they only want to enable it for certain
parts of their app. It leaves control with the app developer.

The `enable_gc_instrumentation` config option should remain so that we
fetch the Garbage Collection profiler wrapper or the nil wrapper
`NilProfiler`.

Part of #868