Add single method to fetch profiler
The MRI probe always initialized the profiler needed to actually profile
the garbage collection. If `enable_gc_instrumentation` is set to
`false`, which it is by default, it has some overhead that's not needed.

Instead it should use the `NilProfiler` like we had in the Transaction
class.

Move this logic to a new `GarbageCollection` wrapper module in which the
`GarbageCollection.profiler` will check the config option value and
return the suitable profiler.

Part of #868