Remove transaction and event level GC tracking
Remove the undocumented transaction and event level Garbage Collection
time tracking. This tracked how much time the GC spent time cleaning up
garbage during the transaction or one of its child events.

This was never shipped, we did not document it or use this data in the
AppSignal.com app. We're unsure about its future in the new Span API and
how to implement this on top of OpenTelemetry spans.

I don't want to have to update the old Transaction API in the extension,
so instead I've changed it to hardcoded `0` values, the same as it
reported with the `GarbageCollection::NilProfiler` by default.

[skip changeset] because it's an undocumented feature I'm removing.