Fix spec relying on Rack environment
In Rack 3.0 the Rack default environment changed to include another key.
We don't really care which values are given by Rack in this test, so use
an `include` matcher to test if some expected keys are at least there.

I've rewritten the test to use the `transaction.to_h` helper so that
testing doesn't rely on stubbing all method calls to the extension.

This won't fix all broken specs, but at least the main AppSignal spec
are running again. That the build is not green yet with this commit is
unfortunately correct.

[skip changeset]