Update CI to Ruby 3.2 rc1
Update Ruby 3.2 prerelease to test if the Ruby gem still works on it.

To make the installation of Ruby 3.2.0-rc1 work on Semaphore I had to
configure RUBY_CFLAGS to "-std=gnu99".

This was mentioned in the bug report for the installation issue:
https://bugs.ruby-lang.org/issues/19161

> From here, I found a simple workaround: ./configure
> cflags="-std=gnu99". Just FYI.

We can probably remove this for the final release, but this way we'll be
able to test against the latest release candidate now.