Add Hanami 2 integration
Hanami 2 is now supported.
To load it, require the instrumentation
file after booting the Hanami 2 application.

Instead of instrumenting Rack requests, instrument calls to the
`call` method in `Hanami::Action` (which receives a Rack request
as an argument) so that we can use the action's class name as the
action name.