Read extension details from Ruby file
Instead of writing a YAML file for every agent release, and then reading
and parsing it on installation, directly write a Ruby file that can be
more easily imported.

This reduces our dependency on YAML library during installation.

Add "stringio" import. That appears to have been loaded by the YAML
require, but now needs to be required manually on Ruby 2.5 and newer.

Part of #904
Closes https://github.com/appsignal/appsignal-agent/issues/736