Fix gem proxy config on Ruby <3.2 with psych 5
We got several reports in issue #904 that installing the AppSignal gem
on system with Ruby < 3.2 with also the psych gem version 5 installed
won't work and raise an error breaking the app installation.

This related PR #913, which is also needed to fix the whole issue. This
PR removes the need for YAML parsing during the installation for the
extension download information.

This change silences the secondary error that is triggered when we fetch
the rubygems config, specified in the `.gemrc` file. This also triggers
that error from within rubygems. Instead, users can use the `HTTP_PROXY`
environment variable to configure the HTTP proxy.

This is not an ideal fix. We also don't communicate that this behavior
is happening with this fix. I suggest we report a bug report upstream as
detailed in #904, so we can remove this rescue-statement.