Add support for GVL metrics
If the GVLTools gem is present, this module adds support for GVL
metrics to be sent to AppSignal. If the GVLTools gem is not present,
the probe that sends the metrics will not start, and the related
configuration options do nothing.

The `enable_gvl_global_timer` and `enable_gvl_waiting_threads`
configuration options can be used to determine whether these
GVLTools instrumentations are enabled when AppSignal starts the
GVL probe. It is possible for the user to disable these config
options and enable GVLTools on demand.

Require GVL version 0.2 or newer, as version 0.2 uses a 64-bit
counter for the instrumentation timers. As it counts in nanoseconds,
the 32-bit counter in previous versions could easily overflow over
a minute, making the delta calculation very unreliable.