Do not require GVLTools on JRuby
The RUBY_VERSION constant refers to the version of the language
implemented by the interpreter, not the version of the interpreter.
These differ for interpreters other than the reference implementation.

Future versions of JRuby will therefore have RUBY_VERSION greater than
or equal to 3.2, but nonetheless they won't be compatible with the
GVLTools library, which is specific to the reference C interpreter.