Fix default Ruby logger formatter usage
The default Ruby logger formatter expects a timestamp as a second
argument to format it and include it in the message. We were passing an
integer instead to the present formatter, making it fail if the present
formatter is the default Ruby one.

This commit substitutes those 0s with `Time.now` calls.