Fix Ruby Logger add method signature
On the previous release, we added a new attribute to the `add` method of
our Logger, which overwrites Ruby's base loggers. This made
ActiveSupport Logger broadcast to fail as we called the function with 4
arguments internally after it was overwrittend by ActiveSupport, which
expects 3 arguments.

This commits addresses the problem by setting the attributes as an
instance variable that is later taken inside the `add` method, so we can
keep the signature untouched.