Report GVL global timer metrics if disabled
We want to document the possibility of enabling and disabling these
metrics on demand. This is why the GVL probe is started even when
`enable_gvl_global_timer` and `enable_gvl_waiting_threads` are both
set to false.

However, this is somewhat undermined by the probe only reporting
values when it is enabled. Meaning, if you only enable it for specific
contexts that you wish to track (for example, for a specific job or
request) and disable it afterwards, the value will only be reported
if the GVL metric happens to be enabled when the probe runs.

There is no actual reason to not report the value that is given by
the GVL global timer, due to its monotonic nature. When disabled, it
will not be incremented, which is fine -- the reported value will be
for the intervals during which it was enabled. Because it can be reset,
do not report the value if the delta goes negative.

The situation is different for the GVL waiting threads. If it is
disabled while threads are waiting, the counter will not decrease for
those threads, essentially raising the baseline for the counter.
Similarly, if it is reset when a thread stops waiting, the counter
can underflow. For now, we won't document disabling the waiting
threads metric, report it only if it's enabled, and hope that it's not
been messed with.