Extract Rails instrumentation metadata earlier
In issue #779, I'm looking at supporting the Rails error reporter. For
this I want to copy the context (transaction namespace, action name and
tags) from the current transaction when an error is reported by Rails.

To have the context available to the error handler, it needs to be set
on the transaction before the controller action or job is performed.
Move those parts of the instrumentation up to above the `super` call,
and not within the `ensure` block at the very end.

Otherwise the context is most likely empty and there is nothing to copy.