Fix Railtie specs usage of the Rails app (#946)
Initialize the Rails app in the Railtie spec. This way we we don't need
to stub calls to the middleware object.

When the app isn't initialized the middleware object is a frozen object
that can't be modified. In a real app, the `app` variable in the Railtie
is also an instance of the app, not the constant.

[skip changeset]