Enable new RuboCop rules and apply fixes
I've disabled the Gemspec/DevelopmentDependencies rule, because I don't
know why we shouldn't add our development dependencies to the gemspec.

If I remove the empty block from the `define_build_task` method, the
tasks for the gem package build are not created. It needs a block
argument, even when empty.