Fix ActiveSupport::TaggedLogging silence usage
When using ActiveSupport::TaggedLogging without the broadcast feature,
any call to `silence` fails.

This patch fixes that behaviour by adding the `silence` method to our
own logger. It acts as the one defined when using `ActiveSupport`'s
broadcast feature. It just calls the passed block and let's the
execution to continue.

This has to be done in order to add full support to
`ActiveSupport::TaggedLogging`.