Renable RuboCop and fix issues
The RuboCop upgrade broke the validation. It only checked 6 files, and
ignored everything else. This is fixed by adding "**/*.rb" to the
included files.

Fix all other issues, mostly automatically corrected.
I updated the to do file for the ones that need someone to take a better
look at them later.

This is not a perfect change. I'm sure we can disagree on some of the
finer points of the style, but let's address that in follow up PRs.
Please review this for any major issues that should not be merged.