Exclude diagnose path from normal RSpec tests
The diagnose tests are in the `spec/integration/diagnose` directory,
which RSpec picks up and tries to run. If I run `rspec` on the command
line, these integration tests are always failing, because they're not
supposed to be run like that, but from within that submodule instead.

Configure RSpec to exclude that directory by default, so it's easier to
run locally without always failing tests.

This isn't an issue on the CI, because we don't checkout the submodule
there, but only for the specific diagnose tests job, so the spec files
aren't available for most jobs.

[skip changeset]