Rails error reporter Sidekiq action and namespace (#955)
Report the Sidekiq worker's action name, namespace and tags for Sidekiq
jobs.

Move the order in which these are set to before the Sidekiq job is
performed. That way that information is available to the Rails error
reporter context and we can copy the action, namespace and tags to the
error reported with `Appsignal.send_error`.

Change the Active Job instrumentation, to use `set_action`, rather than
`set_action_if_nil`. This will overwrite the action name, rather than
only set it if not already set. Because it's set before the job, this
will not overwrite any action name set by the user. This is needed for
the Active Job + Sidekiq instrumentation to work again and use the
Active Job's instrumentation naming for Active Job jobs, run with
Sidekiq.

Part of #779