Refactor Puma spec helper methods
Collect all the Puma plugin state management in the `run_plugin` help,
rather than spread it around the before and after blocks, and
let-statements. This is the result of an attempt to make the test less
fragile/brittle as it fails every once in a while on CI.

Moving everything in the same helper, we're sure what needs to happen
before each start and stop of the plugin and we're sure it always
executes, even on failure.

[skip changeset]