Fix ActiveSupport Logger broadcast compatibility
We were using `progname` instead of group in the AppSignal Logger `add`
method. Making the convenience methods such as `warn`, `info`, etc. Not
logging anything when called as in:

```ruby
Rails.logger.warn("Warning message")
```

This commits updates that to use `group` instead of `progname` as in the
original Ruby Logger implementation.

Co-authored-by: Jörg Schiller <joergschiller@googlemail.com>