Add option to set route definition as Sinatra path
We had a report of an app that contains sensitive information in the
request path and the desire to filter this out. We have no system in
place to filter metadata like path and request method, as set by the
Sinatra middleware.

Add a config option `sinatra_sanitized_routes` that can be set to `true`
to set the route definition rather than the real request path as the
`path` metadata.